refactor: split WASM PAL into focused files
